1.20
- New file format for secondary and IP settings. This version of the firmware  converts old to new.
- "Shuffle" setting changes through buttons are saved (SD required). (Shuffle option removed from CP)
- Superfluous "Brightness" setting removed from CP. Brightness is adjustable using buttons.
- Display MAC address (STA) on WiFi Configuration Page
- Avoid falling back to stand-alone mode too early when TCD stops sending packets
- Config files are now only written if actually changed which prolongs Flash life-span.
- Code optimizations and fixes.

+++ remote-A10001986/display.h
@@ -133,8 +133,8 @@ class remDisplay {
         uint8_t _origBrightness = 15;
 
         int8_t   _dispType = -1;
-        int      _buf_max = 0;      //      number of buffer positions used
-        uint8_t  _num_digs;         //      total number of digits/letters (max 4)
+        unsigned int _buf_max = 0;  //      number of buffer positions used
+        unsigned int _num_digs;     //      total number of digits/letters (max 4)
         const uint8_t *_bufPosArr;  //      Array of buffer positions for digits left->right
         const uint8_t *_bufShftArr; //      Array of shift values for each digit
